# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: API de Finanzas Inteligentes
  version: 1.0.0-draft
  description: Contrato ejecutable derivado de CONVENCIONES.md, ENDPOINTS.md, CONTRATOS.md y TRAZABILIDAD_RF.md.
servers:
- url: https://api.finanzas.example/api/v1
  description: Producción (referencial)
- url: http://localhost:8080/api/v1
  description: Desarrollo local
security:
- bearerAuth: []
...
paths:
...
  "/eventos-seguridad":
    get:
      tags:
      - Seguridad, OTP, biometría y auditoría
      summary: GET /eventos-seguridad
      operationId: getEventosSeguridad
      x-requisitos-funcionales:
      - RF-20
      x-etapa: 1
      x-pruebas-contrato:
      - CT-RF-20
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limite"
      - name: tipo
        in: query
        required: false
        schema:
          type: string
      - name: desde
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: hasta
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PaginaEventoSeguridadResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '500':
          "$ref": "#/components/responses/Problem500"
  "/eventos-auditoria":
    get:
      tags:
      - Seguridad, OTP, biometría y auditoría
      summary: GET /eventos-auditoria
      operationId: getEventosAuditoria
      x-requisitos-funcionales:
      - RF-20
      x-etapa: 1
      x-pruebas-contrato:
      - CT-RF-20
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limite"
      - name: recurso
        in: query
        required: false
        schema:
          type: string
      - name: usuarioId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: grupoFamiliarId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: desde
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: hasta
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PaginaEventoAuditoriaResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '403':
          "$ref": "#/components/responses/Problem403"
        '500':
          "$ref": "#/components/responses/Problem500"
...
  "/dispositivos":
    post:
      tags:
      - Endpoints operativos del cliente
      summary: POST /dispositivos
      operationId: postDispositivos
      x-requisitos-funcionales: []
      x-etapa: 1
      x-pruebas-contrato:
      - CT-DISPOSITIVOS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      responses:
        '201':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            Location:
              "$ref": "#/components/headers/Location"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DispositivoResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '409':
          "$ref": "#/components/responses/Problem409"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-DISPOSITIVOS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostDispositivosRequest"
  "/dispositivos/{dispositivoId}":
    patch:
      tags:
      - Endpoints operativos del cliente
      summary: PATCH /dispositivos/{dispositivoId}
      operationId: patchDispositivosByDispositivoId
      x-requisitos-funcionales: []
      x-etapa: 1
      x-pruebas-contrato:
      - CT-DISPOSITIVOS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: dispositivoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DispositivoResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '412':
          "$ref": "#/components/responses/Problem412"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-DISPOSITIVOS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DispositivoPatchRequest"
    delete:
      tags:
      - Endpoints operativos del cliente
      summary: DELETE /dispositivos/{dispositivoId}
      operationId: deleteDispositivosByDispositivoId
      x-requisitos-funcionales: []
      x-etapa: 1
      x-pruebas-contrato:
      - CT-DISPOSITIVOS
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: dispositivoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '204':
          description: Operación completada sin cuerpo.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '412':
          "$ref": "#/components/responses/Problem412"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-DISPOSITIVOS
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^\".+\"$
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 500
    Limite:
      name: limite
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  headers:
    CorrelationId:
      description: Identificador de correlación de la solicitud.
      schema:
        type: string
    ETag:
      description: Versión opaca del recurso.
      schema:
        type: string
    Location:
      description: URI del recurso creado o aceptado.
      schema:
        type: string
        format: uri-reference
  schemas:
...
    EventoSeguridadResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          example: inicio-sesion
        descripcion:
          type: string
          example: Inicio de sesion desde un dispositivo habitual
        exitoso:
          type: boolean
          example: true
        origenAproximado:
          type: string
          example: Asunción, PY
        dispositivo:
          type: string
          example: Pixel 8
        ocurridoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
      required:
      - id
      - tipo
      - descripcion
      - exitoso
      - origenAproximado
      - dispositivo
      - ocurridoEn
...
    EventoAuditoriaResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        accion:
          type: string
          example: movimiento.creado
        recurso:
          type: string
          example: movimiento
        recursoId:
          type: string
          format: uuid
        usuarioId:
          type: string
          format: uuid
        grupoFamiliarId:
          type:
          - string
          - 'null'
        datosAnteriores:
          type:
          - string
          - 'null'
        datosPosteriores:
          type: object
          additionalProperties: false
          properties:
            monto:
              type: integer
              format: int64
              example: 285000
              minimum: 1
            tipo:
              type: string
              example: gasto
          required:
          - monto
          - tipo
        correlationId:
          type: string
          format: uuid
        ocurridoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
      required:
      - id
      - accion
      - recurso
      - recursoId
      - usuarioId
      - datosPosteriores
      - correlationId
      - ocurridoEn
...
    DispositivoResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        identificadorInstalacion:
          type: string
          example: android-installation-id
        nombre:
          type: string
          example: Pixel 8
        plataforma:
          type: string
          example: android
        versionAplicacion:
          type: string
          example: 0.1.0
        confiable:
          type: boolean
          example: true
        ultimoAccesoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
        version:
          type: integer
          format: int64
          example: 1
      required:
      - id
      - identificadorInstalacion
      - nombre
      - plataforma
      - versionAplicacion
      - confiable
      - ultimoAccesoEn
      - version
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - codigo
      - correlationId
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type:
          - string
          - 'null'
        instance:
          type:
          - string
          - 'null'
        codigo:
          type: string
          pattern: "^[a-z0-9_]+$"
        correlationId:
          type: string
        errores:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
    Paginacion:
      type: object
      additionalProperties: false
      required:
      - siguienteCursor
      - hayMas
      - limite
      properties:
        siguienteCursor:
          type:
          - string
          - 'null'
        hayMas:
          type: boolean
        limite:
          type: integer
          minimum: 1
          maximum: 100
...
    PaginaEventoSeguridadResponse:
      type: object
      additionalProperties: false
      required:
      - datos
      - paginacion
      properties:
        datos:
          type: array
          items:
            "$ref": "#/components/schemas/EventoSeguridadResponse"
        paginacion:
          "$ref": "#/components/schemas/Paginacion"
    PaginaEventoAuditoriaResponse:
      type: object
      additionalProperties: false
      required:
      - datos
      - paginacion
      properties:
        datos:
          type: array
          items:
            "$ref": "#/components/schemas/EventoAuditoriaResponse"
        paginacion:
          "$ref": "#/components/schemas/Paginacion"
...
    PostDispositivosRequest:
      type: object
      additionalProperties: false
      minProperties: 1
    DispositivoPatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre:
          type: string
          example: Pixel 8
        versionAplicacion:
          type: string
          example: 0.1.0
        tokenPush:
          type: string
          writeOnly: true
        zonaHoraria:
          type: string
          example: America/Asuncion
  responses:
    Problem400:
      description: Solicitud o parámetros inválidos.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem401:
      description: Autenticación ausente o inválida.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem403:
      description: Permiso insuficiente.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem404:
      description: Recurso inexistente o no visible.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem409:
      description: Conflicto de estado o duplicado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem412:
      description: ETag desactualizado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem422:
      description: Regla de negocio no satisfecha.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem500:
      description: Error interno no controlado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...